Extract TDG column logic to its own class
